+++ KnockKnock-iOS/Scenes/Feed/FeedWrite/Views/FeedWriteView.swift
@@ -236,17 +236,3 @@ let photoCollectionView = UICollectionView(
     ])
   }
 }
-
-extension UIButton {
-  func alignTextBelow(spacing: CGFloat) {
-    guard let image = self.imageView?.image,
-          let titleLabel = self.titleLabel,
-          let titleText = titleLabel.text else { return }
-
-    let titleSize = titleText.size(withAttributes: [
-      NSAttributedString.Key.font: titleLabel.font as Any
-    ])
-    titleEdgeInsets = UIEdgeInsets(top: spacing, left: -image.size.width, bottom: -image.size.height, right: 0)
-    imageEdgeInsets = UIEdgeInsets(top: -(titleSize.height + spacing), left: 0, bottom: 0, right: -titleSize.width)
-  }
-}
